orders字段默认为null

+++ yunyue-core/yunyue-core-data-mybatis/src/main/java/tech/yunyue/core/mybatis/mvc/entity/mybatis/StdTreeEntity.java
@@ -34,7 +34,7 @@ public class StdTreeEntity extends StdEntity {
     private String parentName;
 
     @Schema(description = "排序")
-    private Integer orders;
+    private Long orders;
     @Schema(description = "路径")
     private String path;
